# MGameInfo-API/war/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.1",
        "title": "MGameInfo",
        "description": "**Find videogames Information!**\n"
    },
    "host": "mgameinfo-api.appspot.com",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/games": {
            "get": {
                "description": "Return a collection of Games.",
                "responses": {
                    "200": {
                        "description": "Games response",
                        "schema": {
                            "$ref": "#/definitions/game"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a game to the collection.",
                "parameters": [
                    {
                        "name": "game",
                        "in": "body",
                        "description": "A Game you want to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Added. Operation successful"
                    }
                }
            }
        },
        "/games/{name}": {
            "get": {
                "description": "Return a game that matches with name.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the game to fetch",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A Game"
                    }
                }
            },
            "put": {
                "description": "Updates an existing Game.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "type": "string",
                        "description": "Name of the Game to Update",
                        "required": true
                    },
                    {
                        "name": "game",
                        "in": "body",
                        "description": "A Game you want to update",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated. Operation successful"
                    }
                }
            },
            "delete": {
                "description": "Deletes an existing Game from the collection.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "type": "string",
                        "description": "A game you want to Delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted. Operation successful"
                    }
                }
            }
        }
    },
    "definitions": {
        "game": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                }
            }
        }
    }
}